day 8: don't divide pitch by gcd
the description/examples are pretty ambiguous
and confusing, especially for part 2. it says

> any grid position exactly in line with
> at least two antennas of the same frequency

but it's not clear what would happen with
a pair of antennas placed such that dx is
coprime to dy. for example, should:
....A...A....
produce these antinodes:
#...#...#...#
or should every cell on that row be
an antinode (since it is exactly in line
with the antennas):
#############

I personally think it's the latest. but this edge
case does not appear in the examples nor in
my input, and the gcd glyph is experimental,
so let's get rid of it